Fix clustering coefficient tests

+++ modules/StatisticsPlugin/src/main/java/org/gephi/statistics/plugin/ClusteringCoefficient.java
@@ -672,7 +672,7 @@ public class ClusteringCoefficient implements Statistics, LongTask {
     }
 
     private void saveCalculatedValue(Node node, Column clusteringColumn,
-            float nodeClusteringCoefficient) {
+            double nodeClusteringCoefficient) {
 
         if (clusteringColumn == null) {
             return;
